made new module/requirements for debug utilites
