+++ requirements-debug.pip
@@ -1,0 +1,1 @@
+memory-profiler==0.21
